Results of students from 2024 Olympiads
The data was scraped as a game in which the students played against the problems. Let p be the score of the student on a particular problem. The following distribution was selected:
• p ≤ 2: the student loses. Often, students can find the first steps, but they have made some progress and are not close to solving the problem. The student is awarded 0 points, and the problem earns 3 − p points.
• 3 ≤ p ≤ 4: the student and the problem have a draw. Three or four points are most often given for considerable progress, and the student is close to solving a problem, but there are still important things that the student has to figure out. Student and problem is awarded 1 point each.
• 5 ≤ p ≤ 7: the student wins. Usually, this happens when a student has solved the problem but made a minor mistake that is easy to fix. The student is awarded p − 4 points, and the problem gets 0 points.

Here we consider three competitions: the National Math Olympiad, the Asian Pacific Math Olympiad (APMO), and the Silk Road Math Com- Competition (SRMC). These three Olympiads are the hardest and most competitive Olympiads held in Kazakhstan: students must be the best in their region to qualify. These three Olympiads are used to identify the best students who will represent Kazakhstan at the International Math Olympiad.
Each Olympiad consists of 6, 5, and 4 problems, respectively. Each student can earn from 0 to 7 points on each of the problems, and the result of the student on the Olympiad is the sum of his points on each of the problems. A student gets 7 points for solving the problem, and 0 points if no progress is made. Partial points are given if the student has made important progress in his work but could not solve the problem. Similarly, points may be deducted if a student has solved the problem, but there are minor mistakes in the student’s solution.
The dataset consists of students from grades 10-11 (12) who have earned non-zero points on the National Olympiad of the respective year. Students in grades 10 and 11 have the same problems, but students from grade 9 have easier ones. Therefore, we have excluded them from the dataset.
It is worth noting that students who have qualified for APMO and SRCM but did not qualify for the National Olympiad achieved low points; therefore, their performance had a negligible effect on the results, and we have excluded them as well for clarity.
Data was taken from the website matol.kz, which contains the results of the students on these three Olympiads in 2024.