All 'Games': Student vs Problem
The performance of each student against each problem is shown.
Format: Date, DateString, index of Team 1, 0, Points Team 1 gained from the game, index of Team 2, 0, Points Team 2 gained from the game
Remark: The data was scraped as a game in which the students played against the problems. Let p be the score of the student on a particular problem. The following distribution was selected:
• p ≤ 2: the student loses. Often, students can find the first steps, but they have made some progress and are not close to solving the problem. The student is awarded 0 points, and the problem earns 3 − p points.
• 3 ≤ p ≤ 4: the student and the problem have a draw. Three or four points are most often given for considerable progress, and the student is close to solving a problem, but there are still important things that the student has to figure out. Student and problem is awarded 1 point each.
• 5 ≤ p ≤ 7: the student wins. Usually, this happens when a student has solved the problem but made a minor mistake that is easy to fix. The student is awarded p − 4 points, and the problem gets 0 points.